chat-rag-services/pyproject.toml
[tool.poetry]
name = "chat-rag-services"
version = "0.1.0"
description = ""
authors = ["Dimas Tri Mustakim <test31122001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.12"
langchain-community = "^0.2.11"
langchain-experimental = "^0.0.64"
python-dotenv = "^1.0.1"
neo4j = "^5.23.1"
langchain-google-vertexai = "^1.0.8"
anthropic = {extras = ["vertex"], version = "^0.32.0"}
tiktoken = "^0.7.0"
fastapi = {extras = ["standard"], version = "^0.112.0"}
langchain-openai = "^0.1.20"
pydantic-settings = "^2.4.0"
python-multipart = "^0.0.9"
requests = "^2.32.3"
neotime = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
